Split TypeScript alias import tests
